# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/org-units": {
            "get": {
                "description": "获取组织节点列表（支持分页、搜索、排序）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "获取组织节点列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10，最大100",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词（搜索组织名称）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式：asc/desc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的组织节点",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "创建组织节点",
                "parameters": [
                    {
                        "description": "组织信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.CreateOrgUnitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_OrgUnitResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            }
        },
        "/org-units/:id": {
            "get": {
                "description": "根据ID获取组织节点详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "获取组织节点详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "组织ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_OrgUnitResponse"
                        }
                    },
                    "404": {
                        "description": "组织不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            },
            "put": {
                "description": "更新组织节点信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "更新组织节点",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "组织ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "组织信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.UpdateOrgUnitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_OrgUnitResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "404": {
                        "description": "组织不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除组织节点",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "删除组织节点",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "组织ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_lyj404_gin-api-template_domain_dto.CreateOrgUnitRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_dto.OrgUnitResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "github_com_lyj404_gin-api-template_domain_dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据列表"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                },
                "total_page": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_dto.UpdateOrgUnitRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_OrgUnitResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.OrgUnitResponse"
                        }
                    ]
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.PaginationResponse"
                        }
                    ]
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-string": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "type": "string"
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
    }
}